Refactor app structure by moving classes to separate files
---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/dreichner2/AIDM?shareId=XXXX-XXXX-XXXX-XXXX).
