+++ aidm_client/__init__.py
@@ -1,0 +1,1 @@
+# Initialize the aidm_client package
