+++ aidm_client/dialogs/__init__.py
@@ -1,0 +1,1 @@
+# Initialize the `dialogs` package
