+++ aidm_client/pages/__init__.py
@@ -1,0 +1,1 @@
+# Initialize the `pages` package
